src/display/requirements.txt
numpy
Pillow
redis
requests
opencv-python
matplotlib
psutil
adafruit-blinka
adafruit-circuitpython-rgb-display
RPi.GPIO
